requirements.txt
markdown
django
django-filter
django-allauth
dj-rest-auth
djangorestframework
djangorestframework-simplejwt